ApplicationCreation
Moves source files to file share, then creates an application\applications based on the information provided by the .csv file. Then creates a folder underneath the Applications based on the Vendor's name for structure, lastly it add the catalog category specified in the csv. For CSV look at ApplicationCreation.csv in the repo.