Create Get-DellCAB.ps1
This script extracts Dell .CAB files and places them in a folder structure used by the ImportDrivers.ps1 from Cortech, which automates the importation of drivers into System Center Configuration Manager. With this script you could download all the .CAB files you need and then Pipe it to this function, see the examples for more details.